requirements.txt
flask==3.1.0
torch==2.5.1
transformers==4.47.1
numpy==2.0.2
pandas==2.2.3
openai-whisper==20240930
pydub==0.25.1
librosa==0.10.2.post1
soundfile==0.12.1
tqdm==4.67.1
